moved commit_block to hotstuff/protocol.rs
