+++ src/state/block_tree.rs
@@ -195,100 +195,6 @@ impl<K: KVStore> BlockTree<K> {
         Ok(())
     }
 
-    /* ↓↓↓ For committing a block in insert_block ↓↓↓ */
-
-    /// Commits a block and its ancestors if they have not been committed already. 
-    /// 
-    /// Returns the hashes of the newly committed blocks, with the updates they caused to the validator set,
-    /// in sequence (from lowest height to highest height).
-    pub fn commit_block(
-        &mut self,
-        wb: &mut BlockTreeWriteBatch<K::WriteBatch>,
-        block: &CryptoHash,
-    ) -> Result<Vec<(CryptoHash, Option<ValidatorSetUpdates>)>, BlockTreeError> {
-
-        // Obtain an iterator over the "block" and its ancestors, all the way until genesis, from newest ("block") to oldest.
-        let blocks_iter =
-            successors(
-            Some(*block), 
-            |b| 
-                self.block_justify(b).ok()
-                .map(|qc| if !qc.is_genesis_qc() {Some(qc.block)} else {None})
-                .flatten()
-            );
-
-        // Newest committed block height, we do not consider the blocks from this height downwards.
-        let min_height = self.highest_committed_block_height()?;
-
-        // Obtain an iterator over the uncomitted blocks among "block" and its ancestors from oldest to newest,
-        // the newest block being "block".
-        // This is required because we want to commit blocks in correct order, applying updates from oldest to
-        // newest.
-        let uncommitted_blocks_iter = 
-            blocks_iter.take_while(|b| min_height.is_none() || min_height.is_some_and(|h| self.block_height(b).unwrap().unwrap() > h));
-        let uncommitted_blocks = uncommitted_blocks_iter.collect::<Vec<CryptoHash>>();
-        let uncommitted_blocks_ordered_iter = uncommitted_blocks.iter().rev();
-        
-        // Helper closure that
-        // (1) commits block b, applying all related updates to the write batch,
-        // (2) extends the vector of blocks committed so far (accumulator) with b together with the optional
-        //     validator set updates associated with b,
-        // (3) returns the extended vector of blocks committed so far (updated accumulator).
-        let commit = 
-            |committed_blocks_res: Result<Vec<(CryptoHash, Option<ValidatorSetUpdates>)>, BlockTreeError>, b: &CryptoHash| 
-            ->  Result<Vec<(CryptoHash, Option<ValidatorSetUpdates>)>, BlockTreeError> 
-        {   
-            let mut committed_blocks = committed_blocks_res?;
-
-            let block_height = self.block_height(b)?.ok_or(BlockTreeError::BlockExpectedButNotFound{block: b.clone()})?;
-            // Work steps:
-
-            // Set block at height.
-            wb.set_block_at_height(block_height, b)?;
-
-            // Delete all of block's siblings.
-            self.delete_siblings(wb, b)?;
-
-            // Apply pending app state updates.
-            if let Some(pending_app_state_updates) = self.pending_app_state_updates(b)? {
-                wb.apply_app_state_updates(&pending_app_state_updates);
-                wb.delete_pending_app_state_updates(b);
-            }
-
-            // Apply pending validator set updates.
-            if let ValidatorSetUpdatesStatus::Pending(validator_set_updates) = self.validator_set_updates_status(b)? {
-
-                let mut committed_validator_set = self.committed_validator_set()?;
-                let previous_validator_set = committed_validator_set.clone();
-                committed_validator_set.apply_updates(&validator_set_updates);
-
-                wb.set_committed_validator_set(&committed_validator_set)?;
-                wb.set_previous_validator_set(&previous_validator_set)?;
-                wb.set_validator_set_update_block_height(block_height)?;
-                wb.set_validator_set_update_completed(false)?;
-                wb.set_committed_validator_set_updates(block)?;
-
-                committed_blocks.push((*b, Some(validator_set_updates.clone())));
-            } else {
-                committed_blocks.push((*b, None));
-            }
-
-            // Update the highest committed block.
-            wb.set_highest_committed_block(b)?;
-
-            // Return the blocks committed so far together with their corresponding validator set updates.
-            Ok(committed_blocks)
-        };
-
-        // Iterate over the uncommitted blocks from oldest to newest, 
-        // (1) applying related updates (by mutating the write batch), and
-        // (2) building up the vector of committed blocks (by pushing the newely committed blocks to
-        //     the accumulator vector).
-        // Finally, return the accumulator.
-        uncommitted_blocks_ordered_iter.fold(Ok(Vec::new()), commit)
-
-    }
-
     /* ↓↓↓ For deleting abandoned branches in insert_block ↓↓↓ */
 
     /// Delete the "siblings" of the specified block, along with all of its associated data (e.g., pending
